remove unused folders/files and reorganize
