+++ src/data-structures/linked-lists/stack-linked-list/index.test.ts
@@ -1,4 +1,4 @@
-import StackLinkedList from ".";
+import StackLinkedList from "./";
 
 const stackLinkedList = new StackLinkedList();
