--- src/data-structures/linked-lists/stack-linked-list/using-functions/index.test.ts
@@ -1,3 +1,0 @@
-describe('Data Structure: Stack Linked List with Function', () => {
-    it.todo('write test cases for StackLinkedList with Function');
-});
